requirements.txt
anyio==3.6.2
bcrypt==4.0.1
click==8.1.3
dnspython==2.3.0
email-validator==1.3.1
fastapi==0.92.0
h11==0.14.0
idna==3.4
psutil==5.9.4
psycopg2==2.9.5
pydantic==1.10.4
PyJWT==2.6.0
sniffio==1.3.0
SQLAlchemy==2.0.4
starlette==0.25.0
typing_extensions==4.5.0
uvicorn==0.20.0
